+++ src/main_window.h
@@ -27,6 +27,7 @@ class MainWindow final : public QMainWindow {
 
 public:
     explicit MainWindow(QWidget* parent = nullptr);
+    ~MainWindow() override;
     void promptForInitialLevel();
 
 protected: